[refactor] containers 안에 gnb, login 폴더 없애기
